chore: WIP library refactor/reorganization, split listener.rs out
