+++ rust/src/lib.rs
@@ -6,7 +6,7 @@ use uuid::Uuid;
 
 use stack_epic_wallet_api::{self, Owner};
 use stack_epic_wallet_config::{WalletConfig, EpicboxConfig};
-use stack_epic_wallet_impls::{DefaultLCProvider, DefaultWalletImpl, EpicboxListenChannel, HTTPNodeClient};
+use stack_epic_wallet_impls::{DefaultLCProvider, DefaultWalletImpl, HTTPNodeClient};
 
 use stack_epic_util::file::get_first_line;
 use stack_epic_wallet_util::epic_util::ZeroingString;
@@ -20,6 +20,16 @@ use stack_epic_util::secp::key::{SecretKey, PublicKey};
 use stack_epic_util::secp::{Secp256k1};
 
 use android_logger::FilterBuilder;
+pub mod config;
+use crate::config::Config;
+use crate::config::create_wallet_config;
+pub mod mnemonic;
+
+use crate::mnemonic::mnemonic;
+use crate::mnemonic::create_seed;
+use crate::mnemonic::_get_mnemonic;
+
+pub mod wallet;
 
 use crate::wallet::Wallet;
 use crate::wallet::open_wallet;
@@ -33,21 +43,14 @@ use crate::wallet::tx_send_http;
 use crate::wallet::WalletInfoFormatted;
 use crate::wallet::nano_to_deci;
 
+pub mod listener;
 
+use crate::listener::Listener;
+use crate::listener::listener_spawn;
+use crate::listener::listener_cancel;
+use crate::listener::listener_cancelled;
 
-
-pub mod config;
-use crate::config::Config;
-use crate::config::create_wallet_config;
-pub mod mnemonic;
-pub mod wallet;
-
-use crate::mnemonic::mnemonic;
-use crate::mnemonic::create_seed;
-use crate::mnemonic::_get_mnemonic;
-
-
-
+#[macro_export]
 macro_rules! ensure_wallet (
     ($wallet_ptr:expr, $wallet:ident) => (
         if ($wallet_ptr as *mut Wallet).as_mut().is_none() {
@@ -73,8 +76,7 @@ extern crate simplelog;
 
 use log::Level;
 use android_logger::Config as AndroidConfig;
-use ffi_helpers::{export_task, Task};
-use ffi_helpers::task::{CancellationToken, TaskHandle};
+use ffi_helpers::task::TaskHandle;
 
 /*
     Create a new wallet
@@ -1338,60 +1340,6 @@ pub fn wallet_scan_outputs(
             return  Err(e);
         }
     }
-}
-
-#[derive(Debug, Clone)]
-pub struct Listener {
-    pub wallet_ptr_str: String,
-    // pub wallet_data: (i64, Option<SecretKey>),
-    pub epicbox_config: String
-}
-
-impl Task for Listener {
-    type Output = usize;
-
-    fn run(&self, cancel_tok: &CancellationToken) -> Result<Self::Output, anyhow::Error> {
-        let mut spins = 0;
-
-        let wallet_data_str = &self.wallet_ptr_str;
-        // let wallet_data = wallet_ptr.to_str().unwrap();
-        let tuple_wallet_data: (i64, Option<SecretKey>) = serde_json::from_str(wallet_data_str).unwrap();
-        let wlt = tuple_wallet_data.0;
-        let sek_key = tuple_wallet_data.1;
-
-        // let wallet_data = &self.wallet_data;
-        // let wlt = wallet_data.clone().0;
-        unsafe {
-            let epicbox_conf = serde_json::from_str::<EpicboxConfig>(&self.epicbox_config.as_str()).unwrap();
-            // let wallet_data = &self.wallet_data;
-            // let wlt = wallet_data.0;
-            // let sek_key = wallet_data.clone().1;
-            ensure_wallet!(wlt, wallet);
-            while !cancel_tok.cancelled() {
-                let listener = EpicboxListenChannel::new().unwrap();
-                let mut reconnections = 0;
-                listener.listen(
-                    wallet.clone(),
-                    Arc::new(Mutex::new(sek_key.clone())),
-                    epicbox_conf.clone(),
-                    &mut reconnections,
-                ).expect("TODO: Error Listening on Epicbox");
-                spins += 1;
-            }
-        }
-        Ok(spins)
-    }
-}
-
-export_task! {
-    Task: Listener;
-    spawn: listener_spawn;
-    wait: listener_wait;
-    poll: listener_poll;
-    cancel: listener_cancel;
-    cancelled: listener_cancelled;
-    handle_destroy: listener_handle_destroy;
-    result_destroy: listener_result_destroy;
 }
 
 #[no_mangle]
